can now label as intended

+++ backend/src/main/scala/TextAudioIndex.scala
@@ -1,6 +1,6 @@
 import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
 
-case class TextAudioIndex(id: Int, samplingRate: Int, textStartPos: Int, textEndPos: Int, audioStartPos: Int, audioEndPos: Int, speakerKey: Int, labeled: Byte, transcriptFileId: Int) {
+case class TextAudioIndex(id: Int, samplingRate: Int, textStartPos: Int, textEndPos: Int, audioStartPos: Double, audioEndPos: Double, speakerKey: Int, labeled: Byte, transcriptFileId: Int) {
 }
 
 object TextAudioIndex {
