+++ backend/src/main/scala/jooq/db/tables/records/TextaudioindexRecord.scala
@@ -5,6 +5,7 @@ package jooq.db.tables.records
 
 
 import java.lang.Byte
+import java.lang.Double
 import java.lang.Integer
 
 import jooq.db.tables.Textaudioindex
@@ -16,7 +17,7 @@ import org.jooq.Row9
 import org.jooq.impl.UpdatableRecordImpl
 
 
-class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Textaudioindex.TEXTAUDIOINDEX) with Record9[Integer, Integer, Integer, Integer, Integer, Integer, Integer, Byte, Integer] {
+class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Textaudioindex.TEXTAUDIOINDEX) with Record9[Integer, Integer, Integer, Integer, Double, Double, Integer, Byte, Integer] {
 
   def setId(value : Integer) : Unit = {
     set(0, value)
@@ -54,22 +55,22 @@ class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Tex
     if (r == null) null else r.asInstanceOf[Integer]
   }
 
-  def setAudiostartpos(value : Integer) : Unit = {
+  def setAudiostartpos(value : Double) : Unit = {
     set(4, value)
   }
 
-  def getAudiostartpos : Integer = {
+  def getAudiostartpos : Double = {
     val r = get(4)
-    if (r == null) null else r.asInstanceOf[Integer]
+    if (r == null) null else r.asInstanceOf[Double]
   }
 
-  def setAudioendpos(value : Integer) : Unit = {
+  def setAudioendpos(value : Double) : Unit = {
     set(5, value)
   }
 
-  def getAudioendpos : Integer = {
+  def getAudioendpos : Double = {
     val r = get(5)
-    if (r == null) null else r.asInstanceOf[Integer]
+    if (r == null) null else r.asInstanceOf[Double]
   }
 
   def setSpeakerkey(value : Integer) : Unit = {
@@ -110,19 +111,19 @@ class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Tex
   // Record9 type implementation
   // -------------------------------------------------------------------------
 
-  override def fieldsRow : Row9[Integer, Integer, Integer, Integer, Integer, Integer, Integer, Byte, Integer] = {
-    super.fieldsRow.asInstanceOf[ Row9[Integer, Integer, Integer, Integer, Integer, Integer, Integer, Byte, Integer] ]
+  override def fieldsRow : Row9[Integer, Integer, Integer, Integer, Double, Double, Integer, Byte, Integer] = {
+    super.fieldsRow.asInstanceOf[ Row9[Integer, Integer, Integer, Integer, Double, Double, Integer, Byte, Integer] ]
   }
 
-  override def valuesRow : Row9[Integer, Integer, Integer, Integer, Integer, Integer, Integer, Byte, Integer] = {
-    super.valuesRow.asInstanceOf[ Row9[Integer, Integer, Integer, Integer, Integer, Integer, Integer, Byte, Integer] ]
+  override def valuesRow : Row9[Integer, Integer, Integer, Integer, Double, Double, Integer, Byte, Integer] = {
+    super.valuesRow.asInstanceOf[ Row9[Integer, Integer, Integer, Integer, Double, Double, Integer, Byte, Integer] ]
   }
   override def field1 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.ID
   override def field2 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.SAMPLINGRATE
   override def field3 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.TEXTSTARTPOS
   override def field4 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.TEXTENDPOS
-  override def field5 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.AUDIOSTARTPOS
-  override def field6 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.AUDIOENDPOS
+  override def field5 : Field[Double] = Textaudioindex.TEXTAUDIOINDEX.AUDIOSTARTPOS
+  override def field6 : Field[Double] = Textaudioindex.TEXTAUDIOINDEX.AUDIOENDPOS
   override def field7 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.SPEAKERKEY
   override def field8 : Field[Byte] = Textaudioindex.TEXTAUDIOINDEX.LABELED
   override def field9 : Field[Integer] = Textaudioindex.TEXTAUDIOINDEX.TRANSCRIPT_FILE_ID
@@ -130,8 +131,8 @@ class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Tex
   override def component2 : Integer = getSamplingrate
   override def component3 : Integer = getTextstartpos
   override def component4 : Integer = getTextendpos
-  override def component5 : Integer = getAudiostartpos
-  override def component6 : Integer = getAudioendpos
+  override def component5 : Double = getAudiostartpos
+  override def component6 : Double = getAudioendpos
   override def component7 : Integer = getSpeakerkey
   override def component8 : Byte = getLabeled
   override def component9 : Integer = getTranscriptFileId
@@ -139,8 +140,8 @@ class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Tex
   override def value2 : Integer = getSamplingrate
   override def value3 : Integer = getTextstartpos
   override def value4 : Integer = getTextendpos
-  override def value5 : Integer = getAudiostartpos
-  override def value6 : Integer = getAudioendpos
+  override def value5 : Double = getAudiostartpos
+  override def value6 : Double = getAudioendpos
   override def value7 : Integer = getSpeakerkey
   override def value8 : Byte = getLabeled
   override def value9 : Integer = getTranscriptFileId
@@ -165,12 +166,12 @@ class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Tex
     this
   }
 
-  override def value5(value : Integer) : TextaudioindexRecord = {
+  override def value5(value : Double) : TextaudioindexRecord = {
     setAudiostartpos(value)
     this
   }
 
-  override def value6(value : Integer) : TextaudioindexRecord = {
+  override def value6(value : Double) : TextaudioindexRecord = {
     setAudioendpos(value)
     this
   }
@@ -190,7 +191,7 @@ class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Tex
     this
   }
 
-  override def values(value1 : Integer, value2 : Integer, value3 : Integer, value4 : Integer, value5 : Integer, value6 : Integer, value7 : Integer, value8 : Byte, value9 : Integer) : TextaudioindexRecord = {
+  override def values(value1 : Integer, value2 : Integer, value3 : Integer, value4 : Integer, value5 : Double, value6 : Double, value7 : Integer, value8 : Byte, value9 : Integer) : TextaudioindexRecord = {
     this.value1(value1)
     this.value2(value2)
     this.value3(value3)
@@ -203,7 +204,7 @@ class TextaudioindexRecord extends UpdatableRecordImpl[TextaudioindexRecord](Tex
     this
   }
 
-  def this(id : Integer, samplingrate : Integer, textstartpos : Integer, textendpos : Integer, audiostartpos : Integer, audioendpos : Integer, speakerkey : Integer, labeled : Byte, transcriptFileId : Integer) = {
+  def this(id : Integer, samplingrate : Integer, textstartpos : Integer, textendpos : Integer, audiostartpos : Double, audioendpos : Double, speakerkey : Integer, labeled : Byte, transcriptFileId : Integer) = {
     this()
 
     set(0, id)
